+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 * Changed Item and Price columns to always be required
 * Fixed an error caused by changing columns before a merchant has been opened
 
+# Version 10.0.7.2-beta
+
+* Added options to select which columns are shown
+* Added merchant columns: Index, ID, Type, and Subtype
+* Fixed errors when trying to look up items before they are cached
+
 # Release History
 
 Full release history can be found [here](https://github.com/kstange/MerchantPlus/wiki/Release-Notes).